centralized mesh function in Python examples
