+++ examples/Python/Basic/mesh_properties.py
@@ -5,23 +5,13 @@
 # examples/Python/Basic/mesh_properties.py
 
 import numpy as np
+import open3d as o3d
 import time
 
-import open3d as o3d
+import meshes
 
 
-def cat_meshes(mesh0, mesh1):
-    mesh = o3d.geometry.TriangleMesh()
-    vertices0 = np.asarray(mesh0.vertices)
-    vertices = np.vstack((vertices0, np.asarray(mesh1.vertices)))
-    mesh.vertices = o3d.utility.Vector3dVector(vertices)
-    triangles = np.vstack((np.asarray(mesh0.triangles),
-                           np.asarray(mesh1.triangles) + vertices0.shape[0]))
-    mesh.triangles = o3d.utility.Vector3iVector(triangles)
-    return mesh
-
-
-def mesh_generator():
+def mesh_generator(edge_cases=True):
     yield 'box', o3d.geometry.create_mesh_box()
     yield 'sphere', o3d.geometry.create_mesh_sphere()
     yield 'cone', o3d.geometry.create_mesh_cone()
@@ -31,47 +21,13 @@ def mesh_generator():
     yield 'moebius (twists=2)', o3d.geometry.create_mesh_moebius(twists=2)
     yield 'moebius (twists=3)', o3d.geometry.create_mesh_moebius(twists=3)
 
-    yield 'knot', o3d.io.read_triangle_mesh('../../TestData/knot.ply')
+    yield 'knot', meshes.knot()
 
-    verts = np.array([[-1, 0, 0], [0, 1, 0], [1, 0, 0], [0, -1, 0], [0, 0, 1]],
-                     dtype=np.float64)
-    triangles = np.array([[0, 1, 3], [1, 2, 3], [1, 3, 4]])
-    mesh = o3d.geometry.TriangleMesh()
-    mesh.vertices = o3d.utility.Vector3dVector(verts)
-    mesh.triangles = o3d.utility.Vector3iVector(triangles)
-    yield 'non-manifold edge', mesh
-
-    verts = np.array([[-1, 0, -1], [1, 0, -1], [0, 1, -1], [0, 0, 0],
-                      [-1, 0, 1], [1, 0, 1], [0, 1, 1]],
-                     dtype=np.float64)
-    triangles = np.array([[0, 1, 2], [0, 1, 3], [1, 2, 3], [2, 0, 3], [4, 5, 6],
-                          [4, 5, 3], [5, 6, 3], [4, 6, 3]])
-    mesh = o3d.geometry.TriangleMesh()
-    mesh.vertices = o3d.utility.Vector3dVector(verts)
-    mesh.triangles = o3d.utility.Vector3iVector(triangles)
-    yield 'non-manifold vertex', mesh
-
-    mesh = o3d.geometry.create_mesh_box()
-    mesh.triangles = o3d.utility.Vector3iVector(np.asarray(mesh.triangles)[:-2])
-    yield 'open box', mesh
-
-    mesh0 = o3d.geometry.create_mesh_box()
-    T = np.eye(4)
-    T[:, 3] += (0.5, 0.5, 0.5, 0)
-    mesh1 = o3d.geometry.create_mesh_box()
-    mesh1.transform(T)
-    mesh = cat_meshes(mesh0, mesh1)
-    yield 'boxes', mesh
-
-
-def edges_to_lineset(mesh, edges, color):
-    ls = o3d.geometry.LineSet()
-    ls.points = mesh.vertices
-    ls.lines = edges
-    colors = np.empty((np.asarray(edges).shape[0], 3))
-    colors[:] = color
-    ls.colors = o3d.utility.Vector3dVector(colors)
-    return ls
+    if edge_cases:
+        yield 'non-manifold edge', meshes.non_manifold_edge()
+        yield 'non-manifold vertex', meshes.non_manifold_vertex()
+        yield 'open box', meshes.open_box()
+        yield 'boxes', meshes.intersecting_boxes()
 
 
 def check_properties(name, mesh):
@@ -99,12 +55,12 @@ def check_properties(name, mesh):
         edges = mesh.get_non_manifold_edges(allow_boundary_edges=True)
         print('  # visualize non-manifold edges (allow_boundary_edges=True)')
         o3d.visualization.draw_geometries(
-            [mesh, edges_to_lineset(mesh, edges, (1, 0, 0))])
+            [mesh, meshes.edges_to_lineset(mesh, edges, (1, 0, 0))])
     if not edge_manifold_boundary:
         edges = mesh.get_non_manifold_edges(allow_boundary_edges=False)
         print('  # visualize non-manifold edges (allow_boundary_edges=False)')
         o3d.visualization.draw_geometries(
-            [mesh, edges_to_lineset(mesh, edges, (0, 1, 0))])
+            [mesh, meshes.edges_to_lineset(mesh, edges, (0, 1, 0))])
     if not vertex_manifold:
         verts = np.asarray(mesh.get_non_manifold_vertices())
         print('  # visualize non-manifold vertices')
@@ -127,7 +83,7 @@ def check_properties(name, mesh):
         edges = np.hstack(edges).T
         edges = o3d.utility.Vector2iVector(edges)
         o3d.visualization.draw_geometries(
-            [mesh, edges_to_lineset(mesh, edges, (1, 1, 0))])
+            [mesh, meshes.edges_to_lineset(mesh, edges, (1, 1, 0))])
     if watertight:
         print('  # visualize watertight mesh')
         o3d.visualization.draw_geometries([mesh])
@@ -146,14 +102,14 @@ if __name__ == "__main__":
     print('#' * 80)
     print('Test mesh properties')
     print('#' * 80)
-    for name, mesh in mesh_generator():
+    for name, mesh in mesh_generator(edge_cases=True):
         check_properties(name, mesh)
 
     # fix triangle orientation
     print('#' * 80)
     print('Fix triangle orientation')
     print('#' * 80)
-    for name, mesh in mesh_generator():
+    for name, mesh in mesh_generator(edge_cases=False):
         mesh.compute_vertex_normals()
         triangles = np.asarray(mesh.triangles)
         rnd_idx = np.random.rand(*triangles.shape).argsort(axis=1)
