+++ examples/Python/Basic/mesh_subdivision.py
@@ -7,31 +7,12 @@
 import numpy as np
 import open3d as o3d
 
-
-def create_mesh_triangle():
-    mesh = o3d.geometry.TriangleMesh()
-    mesh.vertices = o3d.utility.Vector3dVector(
-        np.array([(np.sqrt(8 / 9), 0, -1 / 3),
-                  (-np.sqrt(2 / 9), np.sqrt(2 / 3), -1 / 3),
-                  (-np.sqrt(2 / 9), -np.sqrt(2 / 3), -1 / 3)],
-                 dtype=np.float32))
-    mesh.triangles = o3d.utility.Vector3iVector(np.array([[0, 1, 2]]))
-    return mesh
-
-
-def create_mesh_plane():
-    mesh = o3d.geometry.TriangleMesh()
-    mesh.vertices = o3d.utility.Vector3dVector(
-        np.array([[0, 0, 0], [0, 1, 0], [1, 1, 0], [1, 0, 0]],
-                 dtype=np.float32))
-    mesh.triangles = o3d.utility.Vector3iVector(np.array([[0, 2, 1], [2, 0,
-                                                                      3]]))
-    return mesh
+import meshes
 
 
 def mesh_generator():
-    yield create_mesh_triangle()
-    yield create_mesh_plane()
+    yield meshes.triangle()
+    yield meshes.plane()
     yield o3d.geometry.create_mesh_tetrahedron()
     yield o3d.geometry.create_mesh_box()
     yield o3d.geometry.create_mesh_octahedron()
@@ -39,8 +20,8 @@ def mesh_generator():
     yield o3d.geometry.create_mesh_sphere()
     yield o3d.geometry.create_mesh_cone()
     yield o3d.geometry.create_mesh_cylinder()
-    yield o3d.io.read_triangle_mesh("../../TestData/knot.ply")
-    yield o3d.io.read_triangle_mesh("../../TestData/bathtub_0154.ply")
+    yield meshes.knot()
+    yield meshes.bathtub()
 
 
 if __name__ == "__main__":
