+++ examples/Python/Basic/meshes.py
@@ -1,0 +1,134 @@
+# Open3selfopen3d.org
+# The MIT License (MIT)
+# See license file or visit www.open3d.org for details
+
+# examples/Python/Basic/meshes.py
+
+import numpy as np
+import open3d as o3d
+import os
+import urllib.request
+import gzip
+import tarfile
+import shutil
+
+
+def cat_meshes(mesh0, mesh1):
+    mesh = o3d.geometry.TriangleMesh()
+    vertices0 = np.asarray(mesh0.vertices)
+    vertices = np.vstack((vertices0, np.asarray(mesh1.vertices)))
+    mesh.vertices = o3d.utility.Vector3dVector(vertices)
+    triangles = np.vstack((np.asarray(mesh0.triangles),
+                           np.asarray(mesh1.triangles) + vertices0.shape[0]))
+    mesh.triangles = o3d.utility.Vector3iVector(triangles)
+    return mesh
+
+
+def edges_to_lineset(mesh, edges, color):
+    ls = o3d.geometry.LineSet()
+    ls.points = mesh.vertices
+    ls.lines = edges
+    colors = np.empty((np.asarray(edges).shape[0], 3))
+    colors[:] = color
+    ls.colors = o3d.utility.Vector3dVector(colors)
+    return ls
+
+
+def triangle():
+    mesh = o3d.geometry.TriangleMesh()
+    mesh.vertices = o3d.utility.Vector3dVector(
+        np.array([(np.sqrt(8 / 9), 0, -1 / 3),
+                  (-np.sqrt(2 / 9), np.sqrt(2 / 3), -1 / 3),
+                  (-np.sqrt(2 / 9), -np.sqrt(2 / 3), -1 / 3)],
+                 dtype=np.float32))
+    mesh.triangles = o3d.utility.Vector3iVector(np.array([[0, 1, 2]]))
+    return mesh
+
+
+def plane():
+    mesh = o3d.geometry.TriangleMesh()
+    mesh.vertices = o3d.utility.Vector3dVector(
+        np.array([[0, 0, 0], [0, 0.2, 0], [1, 0.2, 0], [1, 0, 0]],
+                 dtype=np.float32))
+    mesh.triangles = o3d.utility.Vector3iVector(np.array([[0, 2, 1], [2, 0,
+                                                                      3]]))
+    return mesh
+
+
+def non_manifold_edge():
+    verts = np.array([[-1, 0, 0], [0, 1, 0], [1, 0, 0], [0, -1, 0], [0, 0, 1]],
+                     dtype=np.float64)
+    triangles = np.array([[0, 1, 3], [1, 2, 3], [1, 3, 4]])
+    mesh = o3d.geometry.TriangleMesh()
+    mesh.vertices = o3d.utility.Vector3dVector(verts)
+    mesh.triangles = o3d.utility.Vector3iVector(triangles)
+    return mesh
+
+
+def non_manifold_vertex():
+    verts = np.array([[-1, 0, -1], [1, 0, -1], [0, 1, -1], [0, 0, 0],
+                      [-1, 0, 1], [1, 0, 1], [0, 1, 1]],
+                     dtype=np.float64)
+    triangles = np.array([[0, 1, 2], [0, 1, 3], [1, 2, 3], [2, 0, 3], [4, 5, 6],
+                          [4, 5, 3], [5, 6, 3], [4, 6, 3]])
+    mesh = o3d.geometry.TriangleMesh()
+    mesh.vertices = o3d.utility.Vector3dVector(verts)
+    mesh.triangles = o3d.utility.Vector3iVector(triangles)
+    return mesh
+
+
+def open_box():
+    mesh = o3d.geometry.create_mesh_box()
+    mesh.triangles = o3d.utility.Vector3iVector(np.asarray(mesh.triangles)[:-2])
+    return mesh
+
+
+def intersecting_boxes():
+    mesh0 = o3d.geometry.create_mesh_box()
+    T = np.eye(4)
+    T[:, 3] += (0.5, 0.5, 0.5, 0)
+    mesh1 = o3d.geometry.create_mesh_box()
+    mesh1.transform(T)
+    mesh = cat_meshes(mesh0, mesh1)
+    return mesh
+
+
+def knot():
+    mesh = o3d.io.read_triangle_mesh('../../TestData/knot.ply')
+    return mesh
+
+
+def bathtub():
+    mesh = o3d.io.read_triangle_mesh("../../TestData/bathtub_0154.ply")
+    return mesh
+
+
+def armadillo():
+    armadillo_path = '../../TestData/Armadillo.ply'
+    if not os.path.exists(armadillo_path):
+        print('downloading armadillo mesh')
+        url = 'http://graphics.stanford.edu/pub/3Dscanrep/armadillo/Armadillo.ply.gz'
+        urllib.request.urlretrieve(url, armadillo_path + '.gz')
+        print('extract armadillo mesh')
+        with gzip.open(armadillo_path + '.gz', 'rb') as fin:
+            with open(armadillo_path, 'wb') as fout:
+                shutil.copyfileobj(fin, fout)
+        os.remove(armadillo_path + '.gz')
+    return o3d.io.read_triangle_mesh(armadillo_path)
+
+
+def bunny():
+    bunny_path = '../../TestData/Bunny.ply'
+    if not os.path.exists(bunny_path):
+        print('downloading bunny mesh')
+        url = 'http://graphics.stanford.edu/pub/3Dscanrep/bunny.tar.gz'
+        urllib.request.urlretrieve(url, bunny_path + '.tar.gz')
+        print('extract bunny mesh')
+        with tarfile.open(bunny_path + '.tar.gz') as tar:
+            tar.extractall(path=os.path.dirname(bunny_path))
+        shutil.move(
+            os.path.join(os.path.dirname(bunny_path), 'bunny', 'reconstruction',
+                         'bun_zipper.ply'), bunny_path)
+        os.remove(bunny_path + '.tar.gz')
+        shutil.rmtree(os.path.join(os.path.dirname(bunny_path), 'bunny'))
+    return o3d.io.read_triangle_mesh(bunny_path)
